aiiot_gateway/requirements.txt
psutil==5.7.2
pyshark==0.4.2.11
paho-mqtt==1.5.1